// ClientApp/src/Models/FormularioDTO.ts
export class FormularioDTO {
  Nome: string;
  DtNascimento: Date;
  RG: string;
  CPF: string;
  CEP: string;
  Endereco: string;
  NroEndereco: number;
  Complemento: string;
  Bairro: string;
  Cidade: string;
  UF: string;
  Profissao: string;
  TelRes: string;
  TelCel: string;
  AnimaisAdocaoId: number;
}